+++ src-tauri/src/app_error.rs
@@ -5,8 +5,6 @@ use tracing::error;
 pub enum AppError {
     #[error("internal error:")]
     Internal(String),
-    #[error("FS error:")]
-    FS(String),
     #[error("not found")]
     SqlxError(#[from] sqlx::Error),
 }