+++ src-tauri/src/main.rs
@@ -3,128 +3,98 @@
     windows_subsystem = "windows"
 )]
 
-use application_state::ApplicationState;
-use backend_message_handler::{start_message_handler, InternalMessage, WindowVisibility};
 use heartbeat::heartbeat_process;
-use parking_lot::Mutex;
-use std::sync::Arc;
+use message_handler::{MsgI, MsgWV};
+use tauri::generate_context;
 
-#[cfg(debug_assertions)]
-use tauri::Manager;
+use tauri::{Builder, Manager};
+
+use crate::message_handler::MessageHandler;
 
 mod app_error;
 mod application_state;
-mod backend_message_handler;
 mod check_version;
 mod db;
 mod heartbeat;
+mod message_handler;
 mod request_handlers;
-mod system_tray;
-mod window_action;
 
-pub type TauriState<'a> = tauri::State<'a, Arc<Mutex<ApplicationState>>>;
+pub type TauriState<'a> = tauri::State<'a, tokio::sync::broadcast::Sender<MsgI>>;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-enum ObliqoroWindow {
-    Main,
+/// Simple macro to create an empty String, or create String from a &str - to get rid of .to_owned() / String::from() etc
+#[macro_export]
+macro_rules! S {
+    () => {
+        String::new()
+    };
+    ($s:expr) => {
+        String::from($s)
+    };
 }
 
-impl ObliqoroWindow {
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::Main => "main",
-        }
-    }
-}
+const SYSTEM_TRAY_ID: &str = "obliqoro_system_tray";
+const MAIN_WINDOW: &str = "main";
 
 #[tokio::main]
 async fn main() -> Result<(), ()> {
-    let tray = system_tray::create_system_tray();
-
     let (sx, rx) = tokio::sync::broadcast::channel(128);
+    let (sx1, sx2, sx3) = (sx.clone(), sx.clone(), sx.clone());
 
-    let ctx = tauri::generate_context!();
+    let (setup_tx, setup_rx) = tokio::sync::oneshot::channel();
+    MessageHandler::init(rx, sx.clone(), setup_rx);
+    heartbeat_process(&sx);
 
-    match ApplicationState::new(tauri::api::path::app_data_dir(ctx.config()), &sx).await {
-        Err(e) => {
-            tracing::error!("{e:?}");
-            std::process::exit(1);
-        }
-        Ok(app_state) => {
-            // TODO change this to just an Arc<ApplicationState>, and use a message bus everywhere?
-            // Application state could just be an Arc<Sx<InternalMessage>
-            let state = Arc::new(Mutex::new(app_state));
-            let (init_state, internal_state) = (Arc::clone(&state), Arc::clone(&state));
-            let (event_sx, close_sx, handler_sx, tray_sx, instance_sx) =
-                (sx.clone(), sx.clone(), sx.clone(), sx.clone(), sx.clone());
-
-            #[allow(unused_variables)]
-            let app_builder = tauri::Builder::default()
-                .manage(state)
-                .setup(|app| {
-                    #[cfg(debug_assertions)]
-                    {
-                        if let Some(main_window) = app.get_window(ObliqoroWindow::Main.as_str()) {
-                            main_window.open_devtools();
-                        }
-                    }
-                    Ok(())
-                })
-                .system_tray(tray)
-                .on_system_tray_event(move |_app, event| {
-                    system_tray::on_system_tray_event(event, &tray_sx);
-                })
-                .on_window_event(move |handler| match handler.event() {
-                    tauri::WindowEvent::CloseRequested { api, .. } => {
-                        api.prevent_close();
-                        event_sx
-                            .send(InternalMessage::Window(WindowVisibility::Hide))
-                            .ok();
-                    }
-                    tauri::WindowEvent::Moved(val) => {
-                        if val.x <= -32000 && val.y <= -32000 {
-                            event_sx
-                                .send(InternalMessage::Window(WindowVisibility::Minimize))
-                                .ok();
-                        }
-                    }
-                    _ => (),
-                })
-                .invoke_handler(tauri::generate_handler![
-                    request_handlers::init,
-                    request_handlers::minimize,
-                    request_handlers::open_database_location,
-                    request_handlers::pause_after_break,
-                    request_handlers::reset_settings,
-                    request_handlers::set_settings,
-                    request_handlers::toggle_pause,
-                ])
-                .plugin(tauri_plugin_single_instance::init(move |app, argv, cwd| {
-                    instance_sx
-                        .send(InternalMessage::Window(WindowVisibility::Show))
-                        .ok();
-                }))
-                .build(tauri::generate_context!());
-
-            match app_builder {
-                Ok(app) => {
-                    heartbeat_process(&init_state);
-                    start_message_handler(&app, internal_state, rx, handler_sx);
-                    app.run(move |_app, event| {
-                        if let tauri::RunEvent::ExitRequested { api, .. } = event {
-                            close_sx
-                                .send(InternalMessage::Window(WindowVisibility::Hide))
-                                .ok();
-                            api.prevent_exit();
-                        }
-                    });
-                }
-                Err(e) => {
-                    tracing::error!("{:?}", e);
-                    tracing::error!("Unable to build application");
+    Builder::default()
+        .setup(|app| {
+            #[cfg(debug_assertions)]
+            {
+                if let Some(main_window) = app.get_webview_window(MAIN_WINDOW) {
+                    main_window.open_devtools();
                 }
             }
-        }
-    }
+
+            let Ok(app_data_dir) = tauri::path::PathResolver::app_data_dir(app.path()) else {
+                println!("tauri::path::PathResolver::app_data_dir err");
+                std::process::exit(1)
+            };
+            let Ok(system_tray_menu) = application_state::create_system_tray(app.app_handle(), sx)
+            else {
+                println!("create_system_tray err");
+                std::process::exit(1)
+            };
+            setup_tx
+                .send((app.app_handle().to_owned(), app_data_dir, system_tray_menu))
+                .ok();
+            app.manage(sx1);
+            Ok(())
+        })
+        .on_window_event(move |_window, event| match event {
+            tauri::WindowEvent::CloseRequested { api, .. } => {
+                api.prevent_close();
+                sx2.send(MsgI::Window(MsgWV::Hide)).ok();
+            }
+            tauri::WindowEvent::Moved(val) => {
+                if val.x <= -32000 && val.y <= -32000 {
+                    sx2.send(MsgI::Window(MsgWV::Minimize)).ok();
+                }
+            }
+            _ => (),
+        })
+        .invoke_handler(tauri::generate_handler![
+            request_handlers::init,
+            request_handlers::minimize,
+            request_handlers::open_database_location,
+            request_handlers::pause_after_break,
+            request_handlers::reset_settings,
+            request_handlers::set_settings,
+            request_handlers::toggle_pause,
+        ])
+        .plugin(tauri_plugin_single_instance::init(
+            move |_app, _argv, _cwd| {
+                sx3.send(MsgI::Window(MsgWV::Show)).ok();
+            },
+        ))
+        .run(generate_context!())
+        .ok();
     Ok(())
 }