+++ src/Views/TimerView.vue
@@ -51,11 +51,11 @@
 </template>
 
 <script setup lang="ts">
-import { computed } from 'vue';
 import { FrontEndRoutes, InvokeMessage } from '@/types';
 import { sec_to_minutes } from '@/vanillaTS/helpers';
-import { invoke } from '@tauri-apps/api/tauri';
+import { invoke } from '@tauri-apps/api/core';
 import { snackError } from '@/services/snack';
+
 const settingStore = settingModule();
 
 const store = intervalModule();
@@ -87,7 +87,7 @@ watch(auto_resume, async () => {
 
 const current_state = computed(() => settingStore.get_current_state);
 
-// TODO: refactor this, as it's used in the SettingsView as well 
+// TODO: refactor this, as it's used in the SettingsView as well
 const saveTimeout = ref(0);
 
 const send_state = async (): Promise<void> => {
@@ -111,7 +111,7 @@ watch(pauseAfterBreak, async (pause) => {
 });
 
 watch(interval, async (i) => {
-	if (i <= 0) router.push(FrontEndRoutes.Settings);
+	if (i <= 0) await router.push(FrontEndRoutes.Settings);
 });
 
 </script>