+++ src/vanillaTS/helpers.ts
@@ -1,5 +1,5 @@
 
-export const zeroPad = (unit: number ): string => String(unit).padStart(2, '0');
+export const zeroPad = (unit: number): string => String(unit).padStart(2, '0');
 
 export const sec_to_minutes = (s: number): string => {
 	const second = Math.trunc(s % 60);
@@ -11,11 +11,13 @@ export const sec_to_minutes = (s: number): string => {
 	}
 };
 
-/// Work out the correct plural for a given unit
+// Work out the correct plural for a given unit
 const suffix = (unit: number): string => unit === 1 ? '' : 's';
 
-/// Convert from seconds to H, M, S
-/// With an option to hide seconds - used my the session length section
+/*
+ * Convert from seconds to H, M, S
+ * With an option to hide seconds - used my the session length section
+ */
 export const secondsToText = (s: number, hide_second: boolean): string => {
 	const second = Math.trunc(s % 60);
 	const second_string = hide_second ? '' : `, ${zeroPad(second)} seconds`;
@@ -27,8 +29,8 @@ export const secondsToText = (s: number, hide_second: boolean): string => {
 };
 
 
-/// Convert from a percentage to a zeroPadded 2.d.p string
-export const formatPercentage =  (value: number): string => {
+// Convert from a percentage to a zeroPadded 2.d.p string
+export const formatPercentage = (value: number): string => {
 	const rounded = value.toFixed(2);
 	const parts = rounded.split('.');
 	parts[0] = parts[0].padStart(2, '0');